10. tiktok-influencer-creator-audit-tool/requirements.txt
gspread
google-auth
apify-client
pandas